+++ dev/src/views/ui/grid.vue
@@ -23,7 +23,10 @@
       :pager="true"
       :auto="true"
       :resizable="true"
-      :virtual="true">
+      :virtual="true"
+      :sumFields="['age', 'month']"
+      :avgFields="['age', 'year']"
+      :sumAllFields="{ age: 23232 }">
     </XGrid>
   </div>
 </template>
@@ -45,11 +48,105 @@
   };
 
   const columns: Ref<GridColumns> = ref([]);
+  setTimeout(() => {
+    columns.value = [
+      { type: 'checkbox', title: '', width: 60, fixed: 'left' },
+      { type: 'seq', title: '序号', width: 60, fixed: 'left' },
+      {
+        field: 'id',
+        title: 'ID'
+      },
+      {
+        field: 'name',
+        title: '姓名',
+        filters: [{ value: null }],
+        cellRender: {
+          name: 'LinkCell',
+          events: {
+            click(value: any, e: any) {
+              console.log(value, e);
+            }
+          }
+        }
+      },
+      {
+        field: 'avatar',
+        title: '头像',
+        showOverflow: false,
+        visible: true
+      },
+      {
+        field: 'sex',
+        title: '性别',
+        filters: [
+          { label: '男', value: 1 },
+          { label: '女', value: 0 }
+        ],
+        formatter: (data: any) => {
+          const { row } = data;
+          const map: Record<number, any> = {
+            1: '男',
+            0: '女'
+          };
+          return map[row.sex];
+        }
+      },
+      {
+        field: 'age',
+        title: '年龄',
+        sortable: true
+      },
+      {
+        field: 'birth',
+        title: '出生日期',
+        children: [
+          { field: 'year', title: '年' },
+          { field: 'month', title: '月' },
+          { field: 'date', title: '日' }
+        ]
+      },
+      {
+        field: 'address',
+        title: '地址',
+        children: [
+          { field: 'province', title: '省' },
+          { field: 'city', title: '市' },
+          { field: 'county', title: '区' }
+        ]
+      },
+      {
+        field: 'salary',
+        title: '工资',
+        formatter: (e: any) => `￥${numberFormat(e.cellValue, '0.00')}`
+      },
+      {
+        field: 'intro',
+        title: '简介'
+      },
+      {
+        field: 'join',
+        title: '入职日期',
+        filters: [{ value: '' }],
+        filterRender: {
+          name: 'DateFilter'
+        }
+      },
+      {
+        field: 'create',
+        title: '创建时间'
+      },
+      {
+        field: 'actions',
+        title: '操作',
+        width: '150px'
+      }
+    ] as GridColumns;
+  }, 100);
 
   setTimeout(() => {
     columns.value = [
-      { type: 'checkbox', title: '', width: 60 },
-      { type: 'seq', title: '序号', width: 60 },
+      { type: 'checkbox', title: '', width: 60, fixed: 'left' },
+      { type: 'seq', title: '序号', width: 60, fixed: 'left' },
       {
         field: 'id',
         title: 'ID'