+++ packages/ui/src/components/grid/hooks/index.ts
@@ -5,6 +5,5 @@ export * from './useColumnSortable';
 export * from './useCustom';
 export * from './useVxeSlots';
 export * from './useEditRender';
-// export * from './useState';
-// export * from './usePager';
+export * from './useFilterRender';
 export * from './useLoader';