+++ packages/ui/src/components/grid/renderers/components/SelectFilter.vue
@@ -4,16 +4,17 @@
       size="small"
       placeholder="选择筛选项"
       clearable
-      v-bind="mergeProps"
       v-model="state.option.value"
       @change="onChange"
-      @keyup.enter="onKeyup"></SelectEditor>
+      @keyup.enter.stop="onKeyup"
+      v-bind="renderProps"
+      v-on="renderEvents"></SelectEditor>
   </div>
 </template>
 <script lang="ts" setup>
-  import { reactive, computed } from 'vue';
   import SelectEditor from '../../../field/editors/SelectEditor.vue';
   import type { VxeGlobalRendererHandles } from '../../types';
+  import { useFilterRender } from '../../hooks';
 
   export interface Props {
     params: VxeGlobalRendererHandles.RenderFilterParams;
@@ -22,43 +23,8 @@
 
   const props = defineProps<Props>();
 
-  const mergeProps = computed(() => {
-    return {
-      ...props.renderOpts.props,
-      ...props.renderOpts.events
-    };
-  });
-
-  const state = reactive({
-    option: null as any
-  });
-
-  const load = () => {
-    const { params } = props;
-    if (params) {
-      const { column } = params;
-      const option = column.filters[0];
-      state.option = option;
-    }
-  };
-
-  const onChange = () => {
-    const { params } = props;
-    const { option } = state;
-    if (params && option) {
-      const { $panel } = params;
-      const checked = !!option.value;
-      $panel.changeOption(null, checked, option);
-    }
-  };
-
-  const onKeyup = (e: any) => {
-    const { params } = props;
-    if (params) {
-      const { $panel } = params;
-      $panel.confirmFilter(e);
-    }
-  };
+  const { renderProps, renderEvents, state, load, onChange, onKeyup } =
+    useFilterRender(props.renderOpts, props.params);
 
   load();