+++ create-vtj/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.11](https://gitee.com/newgateway/vtj/compare/create-vtj@0.8.10...create-vtj@0.8.11) (2024-07-03)
+
+**Note:** Version bump only for package create-vtj
+
+
+
+
+
+## [0.8.10](https://gitee.com/newgateway/vtj/compare/create-vtj@0.8.9...create-vtj@0.8.10) (2024-06-27)
+
+**Note:** Version bump only for package create-vtj
+
+
+
+
+
 ## [0.8.9](https://gitee.com/newgateway/vtj/compare/create-vtj@0.8.8...create-vtj@0.8.9) (2024-06-06)
 
 **Note:** Version bump only for package create-vtj