+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.17](https://gitee.com/newgateway/vtj/compare/@vtj/cli@0.8.16...@vtj/cli@0.8.17) (2024-07-03)
+
+**Note:** Version bump only for package @vtj/cli
+
+
+
+
+
+## [0.8.16](https://gitee.com/newgateway/vtj/compare/@vtj/cli@0.8.15...@vtj/cli@0.8.16) (2024-06-28)
+
+**Note:** Version bump only for package @vtj/cli
+
+
+
+
+
+## [0.8.15](https://gitee.com/newgateway/vtj/compare/@vtj/cli@0.8.14...@vtj/cli@0.8.15) (2024-06-27)
+
+**Note:** Version bump only for package @vtj/cli
+
+
+
+
+
 ## [0.8.14](https://gitee.com/newgateway/vtj/compare/@vtj/cli@0.8.13...@vtj/cli@0.8.14) (2024-06-06)